Renamed card generator to vanilla card creator
