+++ card_creator/vanilla/generate.bat
