+++ card_creator/vanilla/generate.sh
