+++ card_creator/vanilla/generator.js
