+++ card_creator/vanilla/setup.bat
