+++ card_creator/vanilla/setup.sh
