+++ index.js
@@ -40,14 +40,14 @@ function devmode() {
             cls();
 
             if (vanilla) {
-                if (!fs.existsSync("./card_creator/generator/.ignore.cards.json")) {
+                if (!fs.existsSync("./card_creator/vanilla/.ignore.cards.json")) {
                     watermark();
 
-                    rl.question("Cards file not found! Go to 'card_creator/generator' and run either 'generate.bat' or 'generate.sh', then try again.\n");
+                    rl.question("Cards file not found! Go to 'card_creator/vanilla' and run either 'generate.bat' or 'generate.sh', then try again.\n");
                     continue;
                 }
 
-                require("./card_creator/generator/index").main("./card_creator/generator");
+                require("./card_creator/vanilla/index").main("./card_creator/vanilla");
             } else {
                 require("./card_creator/index").main();
             }
